Feat: "인가 & 인증 설정"
URL별 권한에 따른 인가 설정을 하였음.

1. 토큰이 필요한 것(로그인)
- 게시글 읽기(o)
- qna 게시판 쓰기 (o)
2. 토큰이 필요 && role
- 공지사항 게시판 쓰기(o)
3. 토큰이 필요 && loginId와 일치
- 게시글 수정, 삭제(x)
4. 토큰이 필요 없는 것
- 게시판 list (o)
- 회원가입 (o)
- 로그인 (o)

게시글 수정, 삭제는 토큰이 필요하며 유저들의 pk값으로도 사용가능한 loginId를 게시글과 비교하여 수정, 삭제를 하여야 한다. 아직 개발하지는 못했다.

+++ backend/src/main/java/com/jungwoo/apiserver/domain/Board.java
@@ -1,7 +1,6 @@
 package com.jungwoo.apiserver.domain;
 
 import lombok.*;
-import org.hibernate.annotations.ColumnDefault;
 
 import javax.persistence.*;
 
@@ -25,7 +24,7 @@ public class Board extends BaseTimeEntity{
   private String title;
   private String content;
   private String type;
-  private Long hit;
+  private Integer hit;
 
   private boolean available;
 
@@ -35,7 +34,7 @@ public class Board extends BaseTimeEntity{
 
 
 
-  public void plusViewNum(Long hit) {
+  public void plusViewNum(Integer hit) {
     this.hit = hit+1;
   }
 
@@ -47,12 +46,5 @@ public class Board extends BaseTimeEntity{
   public void changeAvailableBoard(boolean flag){
     this.available = flag;
   }
-
-
-//
-//  @PrePersist
-//  public void prePersist(){
-//    this.available = true;
-//  }
 
 }
